% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNFcall.R
\name{CNF_call}
\alias{CNF_call}
\title{CNA features calling function}
\usage{
CNF_call(data, hg = "hg38")
}
\arguments{
\item{data}{data.frame with colnames: "chromosome", "start", "end", "segVal","sample"}

\item{hg}{the ref genome, support hg38 only.}
}
\value{
CNA features
}
\description{
CNA features calling function
}
\examples{
\donttest{
exampleSeg <- readRDS(system.file("extdata", "exampleSeg.rds",package = "CNAfingerprint", mustWork = TRUE))
features <- CNF_call(exampleSeg,hg="hg38")
}

}
